+++ glomeruli_segmentation/api_interface.py
@@ -1,54 +1,71 @@
 from io import BytesIO
 from logging import Logger
-from typing import Type, Union
+from typing import Tuple, Type, Union
 
 import desert
-import requests
+from aiohttp import ClientResponse, ClientSession, TCPConnector
 from marshmallow import EXCLUDE
 from PIL import Image
-from requests import Response
 
+from glomeruli_segmentation.aiohttp_hooks import get_logging_hooks
 from glomeruli_segmentation.data_classes import Rectangle, Tile, Wsi
 from glomeruli_segmentation.logging_tools import get_logger
-from glomeruli_segmentation.request_hooks import check_for_errors_hook, response_logging_hook
 
 API_VERSION = "v0"
 
 
+class LoggingClientSession(ClientSession):
+    def __init__(self, logger: dict, get_hooks: Tuple = get_logging_hooks, **kwargs):
+        super().__init__(**kwargs)
+        self.hooks = {"response": [get_hook(logger) for get_hook in get_hooks]}
+
+    async def _request(self, method, str_or_url, **kwargs):
+        r = await super()._request(method, str_or_url, **kwargs)
+        for hook in self.hooks["response"]:
+            await hook(r)
+        return r
+
+
+# TODO: add methods for getting from /configuration and post/putting to /failure
 class ApiInterface:
     def __init__(self, api_url: str, job_id: str, headers: dict, logger: Logger = get_logger()):
-
         self.logger = logger
         self.api_url = api_url
         self.job_id = job_id
         self.headers = headers
+        self.session: LoggingClientSession = None
 
-        self.session = requests.Session()
-        request_hooks = [check_for_errors_hook, response_logging_hook]
-        self.session.hooks["response"] = [hook(self.logger) for hook in request_hooks]
+    async def __aenter__(self):
+        self.session = LoggingClientSession(
+            connector=TCPConnector(keepalive_timeout=5, ssl=False, limit=10),
+            headers=self.headers,
+            logger=self.logger,
+        )
+        await self.session.__aenter__()
+        return self
 
-    def get_input(self, key: str, data_class: Type) -> Union[Rectangle, Wsi]:
+    async def __aexit__(self, exc_type, exc_val, exc_tb):
+        if self.session:
+            return await self.session.__aexit__(exc_type, exc_val, exc_tb)
+
+    async def get_input(self, key: str, input_type: Type) -> Union[Rectangle, Wsi]:
 
         url = f"{self.api_url}/{API_VERSION}/{self.job_id}/inputs/{key}"
-        resp = self.session.get(url, headers=self.headers)
+        r = await self.session.get(url)
 
-        try:
-            schema = desert.schema(data_class, meta={"unknown": EXCLUDE})
-            return schema.load(resp.json())
-        except:
-            self.logger.error(resp.json())
-            raise
+        schema = desert.schema(input_type, meta={"unknown": EXCLUDE})
+        response = await r.json()
+        return schema.load(response)
 
-    def post_output(self, key: str, data: dict) -> Response:
+    async def post_output(self, key: str, data: dict) -> ClientResponse:
         """
         post output data by key as defined in EAD
         """
         url = f"{self.api_url}/{API_VERSION}/{self.job_id}/outputs/{key}"
-        resp = self.session.post(url, json=data, headers=self.headers)
+        r = await self.session.post(url, json=data)
+        return await r.json()
 
-        return resp.json()
-
-    def get_wsi_tile(self, rect: Rectangle, slide: Wsi) -> Tile:
+    async def get_wsi_tile(self, slide: Wsi, rect: Rectangle) -> Tile:
         """
         get a WSI tile on level 0
 
@@ -60,13 +77,15 @@ class ApiInterface:
 
         url = f"{self.api_url}/{API_VERSION}/{self.job_id}/regions/{slide.id}/level/{rect.level}/start/{x}/{y}/size/{rect.width}/{rect.height}"
 
-        resp = self.session.get(url, headers=self.headers)
-        return Tile(image=Image.open(BytesIO(resp.content)), rect=rect)
+        r = await self.session.get(url)
+        content = await r.read()
+        return Tile(image=Image.open(BytesIO(content)), rect=rect)
 
-    def put_finalize(self) -> Response:
+    async def put_finalize(self) -> ClientResponse:
         """
         finalize job, such that no more data can be added and to inform EMPAIA infrastructure about job state
         """
         url = f"{self.api_url}/{API_VERSION}/{self.job_id}/finalize"
 
-        return self.session.put(url, headers=self.headers)
+        r = await self.session.put(url)
+        return await r.json()